+++ src/muse/mca.py
@@ -16,6 +16,7 @@ from typing import (
 
 from xarray import Dataset, zeros_like
 
+from muse.outputs.cache import OutputCache
 from muse.readers import read_initial_market
 from muse.sectors import SECTORS_REGISTERED, AbstractSector
 
@@ -61,10 +62,7 @@ class MCA(object):
                 ),
                 timeslices=settings.timeslices,
             ).sel(region=settings.regions)
-        ).interp(
-            year=settings.time_framework,
-            method=settings.interpolation_mode,
-        )
+        ).interp(year=settings.time_framework, method=settings.interpolation_mode)
 
         market["supply"] = zeros_like(market.exports)
         market["consumption"] = zeros_like(market.exports)
@@ -77,6 +75,9 @@ class MCA(object):
             getLogger(__name__).info(f"Created sector {sector}")
 
         outputs = ofactory(*getattr(settings, "outputs", []))
+        outputs_cache = OutputCache(
+            *getattr(settings, "outputs_cache", []), sectors=sectors
+        )
 
         extras = {
             "foresight",
@@ -88,6 +89,7 @@ class MCA(object):
             "root",
             "plugins",
             "outputs",
+            "outputs_cache",
         }
         global_kw = {
             k: v
@@ -107,6 +109,7 @@ class MCA(object):
             sectors=sectors,
             market=market,
             outputs=outputs,  # type: ignore
+            outputs_cache=outputs_cache,  # type: ignore
             **global_kw,
             **carbon_kw,
         )
@@ -116,6 +119,7 @@ class MCA(object):
         sectors: List[AbstractSector],
         market: Dataset,
         outputs: Optional[Callable[[List[AbstractSector], Dataset], Any]] = None,
+        outputs_cache: Optional[OutputCache] = None,
         time_framework: Sequence[int] = list(range(2010, 2100, 10)),
         equilibrium: bool = True,
         equilibrium_variable: Text = "demand",
@@ -184,6 +188,7 @@ class MCA(object):
         self.carbon_method = CARBON_BUDGET_METHODS[carbon_method]
         self.method_options = method_options
         self.outputs = ofactory() if outputs is None else outputs
+        self.outputs_cache = OutputCache() if outputs_cache is None else outputs_cache
 
     def find_equilibrium(
         self,
@@ -360,6 +365,7 @@ class MCA(object):
             self.outputs(
                 self.market, self.sectors, year=self.time_framework[year_idx]
             )  # type: ignore
+            self.outputs_cache.consolidate_cache(year=self.time_framework[year_idx])
             getLogger(__name__).info(f"Finish simulation year {years[0]}!")
 
     def calibrate_legacy_sectors(self):
@@ -623,10 +629,7 @@ def find_equilibrium(
     )
 
 
-def check_demand_fulfillment(
-    market: Dataset,
-    tol: float,
-) -> bool:
+def check_demand_fulfillment(market: Dataset, tol: float) -> bool:
     """Checks if the supply will fulfill all the demand in the future.
 
     If it does not, it logs a warning.