+++ src/muse/objectives.py
@@ -77,6 +77,7 @@ import xarray as xr
 from mypy_extensions import KwArg
 
 from muse.agents import Agent
+from muse.outputs.cache import cache_quantity
 from muse.registration import registrator
 
 OBJECTIVE_SIGNATURE = Callable[
@@ -184,6 +185,7 @@ def register_objective(function: OBJECTIVE_SIGNATURE):
         if "technology" in result.coords:
             raise RuntimeError("Objective should not return a coordinate 'technology'")
         result.name = function.__name__
+        cache_quantity(**{result.name: result})
         return result
 
     return decorated_objective
@@ -622,12 +624,7 @@ def lifetime_levelized_cost_of_energy(
         years <= agent.forecast_year + nyears,
     )
     production = capacity * fixed_outputs * utilization_factor
-    production = convert_timeslice(
-        production,
-        demand.timeslice,
-        QuantityType.EXTENSIVE,
-    )
-
+    production = convert_timeslice(production, demand.timeslice, QuantityType.EXTENSIVE)
     # raw costs --> make the NPV more negative
     # Cost of installed capacity
     installed_capacity_costs = convert_timeslice(
@@ -798,11 +795,7 @@ def net_present_value(
     ).ffill("year")
 
     production = capacity * fixed_outputs * utilization_factor
-    production = convert_timeslice(
-        production,
-        demand.timeslice,
-        QuantityType.EXTENSIVE,
-    )
+    production = convert_timeslice(production, demand.timeslice, QuantityType.EXTENSIVE)
 
     raw_revenues = (production * prices_non_env * rates).sum(("commodity", "year"))