+++ tests/test_outputs.py
@@ -1,6 +1,7 @@
 """Test saving outputs to file."""
 from pathlib import Path
-from typing import Text
+from typing import Dict, Text
+from unittest.mock import patch
 
 import numpy as np
 import pandas as pd
@@ -389,3 +390,301 @@ def test_finite_resources_in_sim(tmp_path, limits_path):
     mca = MCA.factory(read_settings(tmp_path / "model" / "settings.toml"))
     with raises(FiniteResourceException):
         mca.run()
+
+
+def test_register_output_quantity_cache():
+    from muse.outputs.cache import OUTPUT_QUANTITIES, register_cached_quantity
+
+    @register_cached_quantity
+    def dummy_quantity(*args):
+        pass
+
+    assert OUTPUT_QUANTITIES[dummy_quantity.__name__] == dummy_quantity
+
+
+class TestOutputCache:
+    @patch("pubsub.pub.subscribe")
+    @patch("muse.outputs.sector._factory")
+    def test_init(self, mock_factory, mock_subscribe):
+        from muse.outputs.cache import OutputCache
+
+        param = [dict(quantity="height"), dict(quantity="width")]
+        output_quantities = {q["quantity"]: lambda _: None for q in param}
+        output_quantities["depth"] = lambda _: None
+        topic = "BBC Muse"
+
+        output_cache = OutputCache(
+            *param, output_quantities=output_quantities, topic=topic
+        )
+
+        assert mock_factory.call_count == len(param)
+        mock_subscribe.assert_called_once_with(output_cache.cache, topic)
+
+    @patch("pubsub.pub.subscribe")
+    @patch("muse.outputs.sector._factory")
+    def test_cache(self, mock_factory, mock_subscribe):
+        import xarray as xr
+
+        from muse.outputs.cache import OutputCache
+
+        param = [dict(quantity="height"), dict(quantity="width")]
+        output_quantities = {q["quantity"]: lambda _: None for q in param}
+        output_quantities["depth"] = lambda _: None
+        topic = "BBC Muse"
+
+        output_cache = OutputCache(
+            *param, output_quantities=output_quantities, topic=topic
+        )
+
+        output_cache.cache(dict(height=xr.DataArray(), depth=xr.DataArray()))
+
+        assert len(output_cache.to_save.get("height")) == 1
+        assert len(output_cache.to_save.get("depth", [])) == 0
+
+    @patch("pubsub.pub.subscribe")
+    @patch("muse.outputs.sector._factory")
+    def test_consolidate_cache(self, mock_factory, mock_subscribe):
+        import xarray as xr
+
+        from muse.outputs.cache import OutputCache
+
+        param = [dict(quantity="height"), dict(quantity="width")]
+        output_quantities = {q["quantity"]: lambda _: None for q in param}
+        output_quantities["depth"] = lambda _: None
+        topic = "BBC Muse"
+        year = 2042
+
+        output_cache = OutputCache(
+            *param, output_quantities=output_quantities, topic=topic
+        )
+
+        output_cache.cache(dict(height=xr.DataArray()))
+        output_cache.consolidate_cache(year)
+
+        output_cache.factory["height"].assert_called_once()
+
+
+@patch("pubsub.pub.sendMessage")
+@patch("muse.outputs.cache.match_quantities")
+def test_cache_quantity(mock_match, mock_send):
+    from muse.outputs.cache import CACHE_TOPIC_CHANNEL, cache_quantity
+
+    result = {"mass": 42}
+    mock_match.return_value = result
+
+    cache_quantity(**result)
+    mock_send.assert_called_once_with(CACHE_TOPIC_CHANNEL, data=result)
+    mock_send.reset_mock()
+
+    with raises(ValueError):
+        cache_quantity(function=lambda: None, mass=42)
+
+    with raises(ValueError):
+
+        @cache_quantity
+        def fun():
+            pass
+
+    @cache_quantity(quantity="mass")
+    def fun2():
+        return 42
+
+    fun2()
+    mock_match.assert_called_once_with("mass", 42)
+    mock_send.assert_called_once_with(CACHE_TOPIC_CHANNEL, data=result)
+
+
+def test_match_quantities():
+    import xarray as xr
+
+    from muse.outputs.cache import match_quantities
+
+    q = "mass"
+    da = xr.DataArray(name=q)
+    ds = xr.Dataset({q: da})
+
+    def assert_equal(a: Dict[str, xr.DataArray], b: Dict[str, xr.DataArray]):
+        assert set(a.keys()) == set(b.keys())
+        for k in a:
+            xr.testing.assert_equal(a[k], b[k])
+
+    actual = match_quantities(quantity=q, data=da)
+    assert_equal(actual, {q: da})
+
+    actual = match_quantities(quantity=q, data=ds)
+    assert_equal(actual, {q: da})
+
+    p = "height"
+    ds = xr.Dataset({q: da, p: da, "rubish": da})
+    actual = match_quantities(quantity=[q, p], data=ds)
+    assert_equal(actual, {q: da, p: da})
+
+    actual = match_quantities(quantity=[q, p], data=[da, da])
+    assert_equal(actual, {q: da, p: da})
+
+    with raises(ValueError):
+        match_quantities(
+            quantity=[q, p],
+            data=[
+                da,
+            ],
+        )
+
+    with raises(TypeError):
+        match_quantities(quantity=[q, p], data=42)
+
+
+@patch("muse.outputs.cache.extract_agents_internal")
+def test_extract_agents(mock_extract):
+    from muse.outputs.cache import extract_agents
+
+    sectors = set((1, 2, 3))
+    mock_extract.side_effect = [{s: None} for s in sectors]
+
+    actual = extract_agents(sectors)
+
+    assert set(actual.keys()) == sectors
+    assert set(actual.values()) == set((None,))
+
+
+def test_extract_agents_internal(newcapa_agent, retro_agent):
+    from types import SimpleNamespace
+
+    from muse.outputs.cache import extract_agents_internal
+
+    newcapa_agent.name = "A1"
+    retro_agent.name = "A2"
+    sector = SimpleNamespace(name="IT", agents=[newcapa_agent, retro_agent])
+
+    actual = extract_agents_internal(sector)
+    for agent in [newcapa_agent, retro_agent]:
+        assert agent.uuid in actual
+        assert tuple(actual[agent.uuid].keys()) == (
+            "agent",
+            "category",
+            "sector",
+            "dst_region",
+            "year",
+            "installed",
+        )
+        assert actual[agent.uuid]["agent"] == agent.name
+        assert actual[agent.uuid]["category"] == agent.category
+        assert actual[agent.uuid]["sector"] == "IT"
+        assert actual[agent.uuid]["dst_region"] == agent.region
+
+
+def test_aggregate_cache():
+    import numpy as np
+    import xarray as xr
+    from pandas.testing import assert_frame_equal
+
+    from muse.outputs.cache import _aggregate_cache
+
+    quantity = "heigth"
+
+    a = xr.DataArray(np.ones((3, 4, 5)), name=quantity)
+    b = a.copy()
+    b[0, 0, 0] = 0
+
+    actual = _aggregate_cache(quantity, [a, b])
+    assert_frame_equal(actual, b.to_dataframe().reset_index().astype(float))
+
+    actual = _aggregate_cache(quantity, [b, a])
+    assert_frame_equal(actual, a.to_dataframe().reset_index().astype(float))
+
+    c = a.copy()
+    c.assign_coords(dim_0=c.dim_0.data * 10)
+    dc, da = [da.to_dataframe().reset_index() for da in [c, a]]
+
+    actual = _aggregate_cache(quantity, [c, a])
+    expected = pd.DataFrame.merge(dc, da, how="outer").astype(float)
+    assert_frame_equal(actual, expected)
+
+
+def test_consolidate_quantity(newcapa_agent, retro_agent):
+    from types import SimpleNamespace
+
+    from muse.outputs.cache import consolidate_quantity, extract_agents_internal
+
+    newcapa_agent.name = "A1"
+    retro_agent.name = "A2"
+    newcapa_agent.category = "newcapa"
+    retro_agent.category = "retro"
+    sector = SimpleNamespace(name="IT", agents=[newcapa_agent, retro_agent])
+    agents = extract_agents_internal(sector)
+
+    quantity = "heigth"
+    a = xr.DataArray(
+        np.ones((3, 4, 5)),
+        dims=("agent", "replacement", "asset"),
+        coords={
+            "agent": [
+                newcapa_agent.uuid,
+            ]
+            * 3
+        },
+        name=quantity,
+    )
+    b = a.copy()
+    b[0, 0, 0] = 0
+    b.assign_coords(
+        agent=[
+            retro_agent.uuid,
+        ]
+        * 3
+    )
+
+    actual = consolidate_quantity(quantity, [a, b], agents)
+
+    cols = set(
+        list(agents[retro_agent.uuid].keys())
+        + ["installed", "year", "technology", quantity]
+    )
+    assert set(actual.columns) == cols
+    assert all(actual.year == newcapa_agent.forecast_year)
+    assert all(actual.installed == newcapa_agent.year)
+    assert all(
+        [
+            name in (newcapa_agent.name, retro_agent.name)
+            for name in actual.agent.unique()
+        ]
+    )
+
+
+@patch("muse.outputs.cache.consolidate_quantity")
+def test_output_capacity(mock_consolidate):
+    import xarray as xr
+
+    from muse.outputs.cache import capacity
+
+    cached = [xr.DataArray() for _ in range(3)]
+    agents = {}
+
+    capacity(cached, agents)
+    mock_consolidate.assert_called_once_with("capacity", cached, agents)
+
+
+@patch("muse.outputs.cache.consolidate_quantity")
+def test_output_production(mock_consolidate):
+    import xarray as xr
+
+    from muse.outputs.cache import production
+
+    cached = [xr.DataArray() for _ in range(3)]
+    agents = {}
+
+    production(cached, agents)
+    mock_consolidate.assert_called_once_with("production", cached, agents)
+
+
+@patch("muse.outputs.cache.consolidate_quantity")
+def test_output_lcoe(mock_consolidate):
+    import xarray as xr
+
+    from muse.outputs.cache import lcoe
+
+    cached = [xr.DataArray() for _ in range(3)]
+    agents = {}
+
+    lcoe(cached, agents)
+    mock_consolidate.assert_called_once_with("lcoe", cached, agents)